Fold image_texture class into texture.h header
Resolves #295
