--- src/common/surface_texture.h
@@ -1,43 +1,0 @@
-#ifndef SURFACE_TEXTURE_H
-#define SURFACE_TEXTURE_H
-//==============================================================================================
-// Originally written in 2016 by Peter Shirley <ptrshrl@gmail.com>
-//
-// To the extent possible under law, the author(s) have dedicated all copyright and related and
-// neighboring rights to this software to the public domain worldwide. This software is
-// distributed without any warranty.
-//
-// You should have received a copy (see file COPYING.txt) of the CC0 Public Domain Dedication
-// along with this software. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
-//==============================================================================================
-
-#include "common/rtweekend.h"
-#include "common/texture.h"
-
-
-class image_texture : public texture {
-    public:
-        image_texture() {}
-        image_texture(unsigned char *pixels, int A, int B) : data(pixels), nx(A), ny(B) {}
-        virtual vec3 value(double u, double v, const vec3& p) const;
-
-        unsigned char *data;
-        int nx, ny;
-};
-
-vec3 image_texture::value(double u, double v, const vec3& p) const {
-     auto i = static_cast<int>((  u)*nx);
-     auto j = static_cast<int>((1-v)*ny-0.001);
-
-     if (i < 0) i = 0;
-     if (j < 0) j = 0;
-     if (i > nx-1) i = nx-1;
-     if (j > ny-1) j = ny-1;
-
-     auto r = static_cast<int>(data[3*i + 3*nx*j+0]) / 255.0;
-     auto g = static_cast<int>(data[3*i + 3*nx*j+1]) / 255.0;
-     auto b = static_cast<int>(data[3*i + 3*nx*j+2]) / 255.0;
-     return vec3(r, g, b);
-}
-
-#endif
