+++ src/common/texture.h
@@ -20,6 +20,7 @@ class texture  {
         virtual vec3 value(double u, double v, const vec3& p) const = 0;
 };
 
+
 class constant_texture : public texture {
     public:
         constant_texture() {}
@@ -31,6 +32,7 @@ class constant_texture : public texture {
 
         vec3 color;
 };
+
 
 class checker_texture : public texture {
     public:
@@ -64,5 +66,31 @@ class noise_texture : public texture {
         perlin noise;
         double scale;
 };
+
+
+class image_texture : public texture {
+    public:
+        image_texture() {}
+        image_texture(unsigned char *pixels, int A, int B) : data(pixels), nx(A), ny(B) {}
+
+        virtual vec3 value(double u, double v, const vec3& p) const {
+            auto i = static_cast<int>((  u)*nx);
+            auto j = static_cast<int>((1-v)*ny-0.001);
+
+            if (i < 0) i = 0;
+            if (j < 0) j = 0;
+            if (i > nx-1) i = nx-1;
+            if (j > ny-1) j = ny-1;
+
+            auto r = static_cast<int>(data[3*i + 3*nx*j+0]) / 255.0;
+            auto g = static_cast<int>(data[3*i + 3*nx*j+1]) / 255.0;
+            auto b = static_cast<int>(data[3*i + 3*nx*j+2]) / 255.0;
+            return vec3(r, g, b);
+        }
+
+        unsigned char *data;
+        int nx, ny;
+};
+
 
 #endif
